Fix MSVC warnings

+++ source/test/vector-test.hpp
@@ -12,8 +12,8 @@
         auto rs_unit_test_rhs = rhs; \
         REQUIRE(rs_unit_test_lhs.size() == rs_unit_test_rhs.size()); \
         auto rs_unit_test_epsilon = double(epsilon); \
-        for (size_t i = 0; i < rs_unit_test_lhs.size(); ++i) { \
-            if (std::abs(double(rs_unit_test_rhs[i] - rs_unit_test_lhs[i])) > rs_unit_test_epsilon) { \
+        for (int i = 0; i < int(rs_unit_test_lhs.size()); ++i) { \
+            if (std::abs(double(rs_unit_test_rhs[i]) - double(rs_unit_test_lhs[i])) > rs_unit_test_epsilon) { \
                 FAIL_TEST("Vectors are not close enough: " \
                     << # lhs << " = " << rs_unit_test_lhs << ", " \
                     << # rhs << " = " << rs_unit_test_rhs << ", " \
@@ -38,8 +38,8 @@
         REQUIRE(rs_unit_test_lhs.size() == rs_unit_test_rhs.size()); \
         auto rs_unit_test_epsilon = double(epsilon); \
         bool rs_unit_test_unsaturated = rs_unit_test_lhs[1] < epsilon && rs_unit_test_rhs[1] < epsilon; \
-        for (auto i = size_t(rs_unit_test_unsaturated); i < rs_unit_test_lhs.size(); ++i) { \
-            if (std::abs(double(rs_unit_test_rhs[i] - rs_unit_test_lhs[i])) > rs_unit_test_epsilon) { \
+        for (auto i = int(rs_unit_test_unsaturated); i < int(rs_unit_test_lhs.size()); ++i) { \
+            if (std::abs(double(rs_unit_test_rhs[i]) - double(rs_unit_test_lhs[i])) > rs_unit_test_epsilon) { \
                 FAIL_TEST("Vectors are not close enough: " \
                     << # lhs << " = " << rs_unit_test_lhs << ", " \
                     << # rhs << " = " << rs_unit_test_rhs << ", " \
